isofs: fix timestamps beyond 2027
isofs uses a 'char' variable to load the number of years since
1900 for an inode timestamp. On architectures that use a signed
char type by default, this results in an invalid date for
anything beyond 2027.

This adds a cast to 'u8' for the year number, which should extend
the shelf life of the file system until 2155.


Conflicts:
	fs/isofs/util.c

+++ fs/isofs/util.c
@@ -20,7 +20,7 @@ int iso_date(char * p, int flag)
 	int year, month, day, hour, minute, second, tz;
 	int crtime;
 
-	year = p[0];
+	year = (int)(u8)p[0];
 	month = p[1];
 	day = p[2];
 	hour = p[3];
